requirements.txt
flask
gunicorn
python-binance
pandas
openpyxl
rq